requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/dev.in
#
appdirs==1.4.3            # via black
astroid==2.0.4            # via pylint
attrs==18.2.0             # via black
backcall==0.1.0           # via ipython
black==18.9b0
certifi==2018.11.29       # via requests
chardet==3.0.4            # via requests
click==7.0                # via black
coverage==4.5.2
decorator==4.3.0          # via ipython, traitlets
defusedxml==0.5.0         # via python3-openid
django-allauth==0.39.1
django-braces==1.13.0
django-crispy-forms==1.7.2
django-filter==2.0.0
django-tables2==2.0.3
django==2.1.7             # via django-allauth, django-filter, django-tables2, model-mommy
et-xmlfile==1.0.1         # via openpyxl
filelock==3.0.10          # via tox
flake8==3.6.0
idna==2.8                 # via requests
ipdb==0.11
ipython-genutils==0.2.0   # via traitlets
ipython==7.1.1            # via ipdb
isort==4.3.4
jdcal==1.4                # via openpyxl
jedi==0.13.1              # via ipython
lazy-object-proxy==1.3.1  # via astroid
mccabe==0.6.1             # via flake8, pylint
model-mommy==1.6.0
oauthlib==3.0.1           # via requests-oauthlib
odfpy==1.3.6              # via tablib
openpyxl==2.5.10          # via tablib
parso==0.3.1              # via jedi
pep8==1.7.1
pexpect==4.6.0            # via ipython
pickleshare==0.7.5        # via ipython
pluggy==0.8.0             # via tox
prompt-toolkit==2.0.7     # via ipython
psycopg2-binary==2.7.6.1
ptyprocess==0.6.0         # via pexpect
py==1.7.0                 # via tox
pycodestyle==2.4.0        # via flake8
pyflakes==2.0.0           # via flake8
pygments==2.2.0           # via ipython
pylint-django==2.0.2
pylint-plugin-utils==0.4  # via pylint-django
pylint==2.1.1
python3-openid==3.1.0     # via django-allauth
pytz==2018.7              # via django
pyyaml==4.2b4             # via tablib
requests-oauthlib==1.2.0  # via django-allauth
requests==2.21.0          # via django-allauth, requests-oauthlib
six==1.11.0               # via astroid, model-mommy, prompt-toolkit, tox, traitlets
tablib==0.12.1
toml==0.10.0              # via black, tox
tox==3.5.3
traitlets==4.3.2          # via ipython
typed-ast==1.1.0          # via astroid
unicodecsv==0.14.1        # via tablib
urllib3==1.24.1           # via requests
virtualenv==16.1.0        # via tox
wcwidth==0.1.7            # via prompt-toolkit
wrapt==1.10.11            # via astroid
xlrd==1.1.0               # via tablib
xlwt==1.3.0               # via tablib
yapf==0.24.0
